refactor: extract shared CreateFormKit for skill/agent create pages
CreateSkillPage and CreateAgentPage carried near-identical copies of
ModelPicker, ToolsInput, FieldHint, CharCounter, TopBar, the field
constants, name validation and the save/cancel keyboard handler. Move
them into shared/CreateFormKit, parametrized by accent (accent/violet)
and placeholders, so both pages share one source of truth.
